Compute the maximal partial clique given an adjacency matrix and required density

Compute the maximal partial clique given an adjacency matrix and required density

Usage

compute_maximal_partial_clique_2(adj_mat, alpha)

Arguments

adj_mat

adjacency matrix that is symmetric, only has entries of 0 and 1, has row/column sizes between 5 and 50, has no row/col names, and has 1 across the diagonal

alpha

required density for the partial clique, a number between 0.5 and 1

The method works by beginning with considering the full set of nodes (ordered by degree/connectedness) and then seeing if the edge density requirement (for the given alpha) is achieved. If it isn't, it removes the least connected node and tries again until a group of nodes (partial clique) fulfilling the edeg gensity requirements is found. This is a faster solution but is not perfect, as it can exclude nodes that are indeed part of a grander partial clique early on. Since we are generally specifying high alphas when testing, however, we are assuming that the partial cliques should be quite connected and thus nodes with very little connections often won't be part of the clique.

Value

a list with 1. the index positions of nodes in the maximum partial clique found and 2. the associated edge density for the partial clique