This function runs the simulation.

This function runs the simulation.

Usage

execute_simulation(
  trials_per_level,
  n_nodes,
  clique_fraction,
  alpha,
  seed = 452129
)

Arguments

trials_per_level

A natural number specifying how many times to run each level

n_nodes

A vector of the number of nodes to create a graph with, e.g. c(10) or c(10, 20, 30)

clique_fraction

A vector of the clique fraction

alpha

A vector of alpha values

seed

For setting the seed

Value

A data frame with the simulation results

Examples

library(UWBiost561)
execute_simulation(trials_per_level=1,
                   n_nodes=c(10),
                   clique_fraction=c(0.5),
                   alpha=c(0.6))
#>    number valid    status clique_size run_time n_nodes clique_fraction alpha
#> 1       1  TRUE completed           5    0.004      10             0.5   0.6
#> 2       2  TRUE completed           4    0.006      10             0.5   0.6
#> 3       3  TRUE completed           5    0.003      10             0.5   0.6
#> 4       4  TRUE completed           1    0.001      10             0.5   0.6
#> 5       5  TRUE completed           5    0.003      10             0.5   0.6
#> 6       6  TRUE completed           5    0.002      10             0.5   0.6
#> 7       7 FALSE completed           4    0.001      10             0.5   0.6
#> 8       8  TRUE completed           5    0.001      10             0.5   0.6
#> 9       9  TRUE completed           5    0.002      10             0.5   0.6
#> 10     10  TRUE completed           5    0.002      10             0.5   0.6
#> 11     11  TRUE completed           5    0.059      10             0.5   0.6
#> 12     12  TRUE completed           5    0.002      10             0.5   0.6
#> 13     13  TRUE completed           5    0.003      10             0.5   0.6
#> 14     14  TRUE completed           5    0.001      10             0.5   0.6
#> 15     15 FALSE timed_out           1   29.343      10             0.5   0.6