Run a simulation of maximal partial clique computing implementations

Run a simulation of maximal partial clique computing implementations

Usage

maximal_partial_clique_sim(
  imp_numbers = 1:15,
  trials = 4,
  alpha_vec = c(0.5, 0.75, 0.95),
  n_vec = c(10, 25, 45)
)

Arguments

imp_numbers

number of implementations

trials

number of trials to be ran for each level per implementation

alpha_vec

vector of alpha levels. Alpha is a numeric threshold between 1 and 0.5 indicating required edge density

n_vec

vector of n levels. n is the number of nodes in the graph

Value

a nested list with results for each level, each trial, and each implementation