Generate partial clique Creates a graph with a partial clique according to the values passed. The output is the graph, the index so that the graph can be reordered to make the partial clique more apparent, and lastly, the size and number of edges in the partial clique.

Generate partial clique Creates a graph with a partial clique according to the values passed. The output is the graph, the index so that the graph can be reordered to make the partial clique more apparent, and lastly, the size and number of edges in the partial clique.

Usage

generate_partial_clique(
  n,
  clique_fraction,
  clique_edge_density,
  non_clique_edge_density = 0.1
)

Arguments

n

The size of the matrix to be generated

clique_fraction

the fraction of the nodes should be in the partial clique

clique_edge_density

The density of the edges in the partial clique, excluding edges from a node to itself

non_clique_edge_density

The density of edge not withing the clique, also excluding self-edges

Value

A list containing the adjacency matrix, the index that reverse the permutation applied (making it easier to see the clique), the size of the partial clique and the number of edges it contains